Split the code in index.ts into separate files
