+++ typescript/package.json
@@ -13,9 +13,9 @@
     "typescript": "5.0.4"
   },
   "scripts": {
-    "build": "tsc -m node16 --target es2020 --declaration index.ts esil.ts shell.ts",
+    "build": "tsc -m node16 --target es2020 --declaration r2pipe.ts base64.ts ai.ts r2papi.ts esil.ts shell.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "doc": "typedoc index.ts shell.ts esil.ts"
+    "doc": "typedoc r2pipe.ts ai.ts r2papi.ts base64.ts shell.ts esil.ts"
   },
   "repository": {
     "type": "git",
@@ -34,15 +34,24 @@
   ],
   "files": [
     "README.md",
+    "ai.js",
+    "ai.ts",
+    "ai.d.ts",
+    "base64.js",
+    "base64.ts",
+    "base64.d.js",
+    "r2pipe.js",
+    "r2pipe.ts",
+    "r2pipe.d.ts",
     "esil.js",
     "esil.ts",
     "esil.d.ts",
     "shell.js",
     "shell.ts",
     "shell.d.ts",
-    "index.js",
-    "index.ts",
-    "index.d.ts"
+    "r2papi.js",
+    "r2papi.ts",
+    "r2papi.d.ts"
   ],
   "license": "MIT"
 }
